Remove: Factory중 일부 제거
Factory 중 타입을 생성하는 팩토리들을 제거하고 각 타입에 random() 타입 메서드를 추가
